Hotfix c7 nf 3356

See merge request choerodon-framework/iam-service!428

+++ src/main/java/io/choerodon/iam/domain/service/impl/ParsePermissionServiceImpl.java
@@ -47,8 +47,8 @@ public class ParsePermissionServiceImpl implements ParsePermissionService {
 
     private RoleRepository roleRepository;
 
-    @Value("${choerodon.cleanRolePermission:true}")
-    private boolean cleanErrorRolePermission;
+    @Value("${choerodon.cleanPermission:false}")
+    private boolean cleanPermission;
 
     public void setRestTemplate(RestTemplate restTemplate) {
         this.restTemplate = restTemplate;
@@ -79,16 +79,11 @@ public class ParsePermissionServiceImpl implements ParsePermissionService {
             String serviceName = payload.getAppName();
             String json = payload.getApiData();
             if (logger.isDebugEnabled()) {
-                logger.debug("receive message from manager-service, service: {}, version: {}, ip: {}, ###swagger json###  {}", serviceName, payload.getVersion(), payload.getInstanceAddress(), json);
+                logger.debug("receive service: {} message, version: {}, ip: {}, ###swagger json###  {}", serviceName, payload.getVersion(), payload.getInstanceAddress(), json);
             } else {
-                logger.info("receive message from manager-service, service: {}, version: {}, ip: {}", serviceName, payload.getVersion(), payload.getInstanceAddress());
+                logger.info("receive service: {} message, version: {}, ip: {}", serviceName, payload.getVersion(), payload.getInstanceAddress());
             }
             if (!StringUtils.isEmpty(serviceName) && !StringUtils.isEmpty(json)) {
-                //清理role_permission表层级不符的脏数据，会导致基于角色创建失败
-                if (cleanErrorRolePermission) {
-                    logger.info("begin to clean the error role_permission");
-                    cleanRolePermission();
-                }
                 JsonNode node = objectMapper.readTree(json);
                 Iterator<Map.Entry<String, JsonNode>> pathIterator = node.get("paths").fields();
                 Map<String, RoleDO> initRoleMap = queryInitRoleByCode();
@@ -98,7 +93,11 @@ public class ParsePermissionServiceImpl implements ParsePermissionService {
                     Iterator<Map.Entry<String, JsonNode>> methodIterator = pathNode.getValue().fields();
                     parserMethod(methodIterator, pathNode, serviceName, initRoleMap, permissionCodes);
                 }
-                deleteDeprecatedPermission(permissionCodes, serviceName);
+                if (cleanPermission) {
+                    deleteDeprecatedPermission(permissionCodes, serviceName);
+                    //清理role_permission表层级不符的脏数据，会导致基于角色创建失败
+                    cleanRolePermission();
+                }
             }
         } catch (IOException e) {
             throw new CommonException("error.parsePermissionService.parse.IOException", e);
@@ -130,7 +129,7 @@ public class ParsePermissionServiceImpl implements ParsePermissionService {
             for (PermissionDO permission : permissions) {
                 RolePermissionE rp = new RolePermissionE(null, role.getId(), permission.getId());
                 rolePermissionRepository.delete(rp);
-                logger.debug("delete error role_permission, role id: {}, code: {}, level: {} ## permission id: {}, code:{}, level: {}",
+                logger.info("delete error role_permission, role id: {}, code: {}, level: {} ## permission id: {}, code:{}, level: {}",
                         role.getId(), role.getCode(), role.getLevel(), permission.getId(), permission.getCode(), permission.getLevel());
                 count++;
             }
